Move IValidatorStateService to its own file
